requirements.txt
ebrains-kg-core
fairgraph
ebrains_drive